Move MaybePopup
