+++ shared/teams/really-leave-team/index.js
@@ -1,21 +1,8 @@
 // @flow
 import * as React from 'react'
-import {Avatar, Box, Text, Icon, PopupDialog, Button} from '../../common-adapters'
-import {globalStyles, globalMargins, globalColors} from '../../styles'
+import {Avatar, Box, Text, Icon, Button, MaybePopup} from '../../common-adapters'
+import {globalStyles, globalMargins} from '../../styles'
 import {isMobile} from '../../constants/platform'
-
-const MaybePopup = isMobile
-  ? (props: {onClose: () => void, children: React.Node}) => (
-      <Box style={{height: '100%', width: '100%'}} children={props.children} />
-    )
-  : (props: {onClose: () => void, children: React.Node}) => (
-      <PopupDialog
-        onClose={props.onClose}
-        styleCover={_styleCover}
-        styleContainer={_styleContainer}
-        children={props.children}
-      />
-    )
 
 type Props = {
   onClose: () => void,
@@ -60,13 +47,4 @@ const ReallyLeaveTeam = (props: Props) => (
   </MaybePopup>
 )
 
-const _styleCover = {
-  alignItems: 'stretch',
-  backgroundColor: globalColors.black_75,
-  justifyContent: 'stretch',
-}
-
-const _styleContainer = {
-  height: '100%',
-}
 export default ReallyLeaveTeam
